- Generic update

+++ restfulapi_cinetown/src/main/java/com/execinema/restfulapi_cinetown/service/FilmServiceImpl.java
@@ -4,11 +4,13 @@ package com.execinema.restfulapi_cinetown.service;
 import com.execinema.restfulapi_cinetown.api.model.FilmDTO;
 import com.execinema.restfulapi_cinetown.domain.Film;
 import com.execinema.restfulapi_cinetown.domain.FilmId;
+import com.execinema.restfulapi_cinetown.exception.ResourceNotFoundException;
 import com.execinema.restfulapi_cinetown.repository.FilmRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Optional;
 
 @Service
 public class FilmServiceImpl implements FilmService{
@@ -33,13 +35,21 @@ public class FilmServiceImpl implements FilmService{
     @Override
     public void deleteFilmByNameAndProducer(String filmName, String producer) {
         FilmId filmId = new FilmId(filmName, producer);
+
+        Optional<Film> film = filmRepository.findById(filmId);
+        if(film.isEmpty()){
+            throw new ResourceNotFoundException("Non esiste il fil nel db");
+        }
         filmRepository.deleteById(filmId);
     }
 
 
     @Override
     @Transactional
-    public void deleteFilmByName(String filmName) {
+    public void deleteFilmByName(String filmName){
+
+        //dovrei definire il metodo nella repository per poterlo utilizzare
+        /*Set<Film> findingFilms = filmRepository.findByFilmId_Name*/
         filmRepository.deleteAllByName(filmName);
     }
